# Make Attributesrecord use pointers to reduce copying