--- a/pkg/registry/authorization/util/helpers.go
+++ b/pkg/registry/authorization/util/helpers.go
@@ -23,8 +23,8 @@ import (
 )
 
 // ResourceAttributesFrom combines the API object information and the user.Info from the context to build a full authorizer.AttributesRecord for resource access
-func ResourceAttributesFrom(user user.Info, in authorizationapi.ResourceAttributes) authorizer.AttributesRecord {
-	return authorizer.AttributesRecord{
+func ResourceAttributesFrom(user user.Info, in authorizationapi.ResourceAttributes) *authorizer.AttributesRecord {
+	return &authorizer.AttributesRecord{
 		User:            user,
 		Verb:            in.Verb,
 		Namespace:       in.Namespace,
@@ -38,8 +38,8 @@ func ResourceAttributesFrom(user user.Info, in authorizationapi.ResourceAttribut
 }
 
 // NonResourceAttributesFrom combines the API object information and the user.Info from the context to build a full authorizer.AttributesRecord for non resource access
-func NonResourceAttributesFrom(user user.Info, in authorizationapi.NonResourceAttributes) authorizer.AttributesRecord {
-	return authorizer.AttributesRecord{
+func NonResourceAttributesFrom(user user.Info, in authorizationapi.NonResourceAttributes) *authorizer.AttributesRecord {
+	return &authorizer.AttributesRecord{
 		User:            user,
 		ResourceRequest: false,
 		Path:            in.Path,
@@ -60,7 +60,7 @@ func convertToUserInfoExtra(extra map[string]authorizationapi.ExtraValue) map[st
 }
 
 // AuthorizationAttributesFrom takes a spec and returns the proper authz attributes to check it.
-func AuthorizationAttributesFrom(spec authorizationapi.SubjectAccessReviewSpec) authorizer.AttributesRecord {
+func AuthorizationAttributesFrom(spec authorizationapi.SubjectAccessReviewSpec) *authorizer.AttributesRecord {
 	userToCheck := &user.DefaultInfo{
 		Name:   spec.User,
 		Groups: spec.Groups,
@@ -68,12 +68,9 @@ func AuthorizationAttributesFrom(spec authorizationapi.SubjectAccessReviewSpec) 
 		Extra:  convertToUserInfoExtra(spec.Extra),
 	}
 
-	var authorizationAttributes authorizer.AttributesRecord
 	if spec.ResourceAttributes != nil {
-		authorizationAttributes = ResourceAttributesFrom(userToCheck, *spec.ResourceAttributes)
-	} else {
-		authorizationAttributes = NonResourceAttributesFrom(userToCheck, *spec.NonResourceAttributes)
+		return ResourceAttributesFrom(userToCheck, *spec.ResourceAttributes)
 	}
 
-	return authorizationAttributes
+	return NonResourceAttributesFrom(userToCheck, *spec.NonResourceAttributes)
 }