… plugin/pkg/auth/authorizer/rbac/rbac_test.go
… staging/src/k8s.io/apiserver/pkg/authorization/authorizerfactory/builtin_test.go
… staging/src/k8s.io/apiserver/pkg/endpoints/handlers/responsewriters/errors_test.go
… staging/src/k8s.io/apiserver/plugin/pkg/authorizer/webhook/webhook_test.go
… test/integration/auth/auth_test.go