Change AttributesRecord to use a pointer receiver
Allows to use the structure without copying a bunch of stuff around every time

```release-note
NONE
```