@enj 2018-01-18
/lgtm

@kubernetes/sig-auth-pr-reviews 

@wojtek-t any thoughts on how this could positively or negatively effect performance?  My guess is the copies lived on the stack, and the pointers will now live on the heap.