@deads2k 2018-01-18
@smarterclayton I remember with groupversion and the like it was actually faster to use structs than pointers.  Any idea if this will follow the same pattern?