@liggitt 2018-01-18
no difference in speed or allocations through the RBAC stack:

on master
```
$ go test ./plugin/pkg/auth/authorizer/rbac/ -run foo -bench . -benchmem
goos: darwin
goarch: amd64
pkg: k8s.io/kubernetes/plugin/pkg/auth/authorizer/rbac
BenchmarkAuthorize/allow_list_pods-8         	  500000	      2694 ns/op	    1632 B/op	      27 allocs/op
BenchmarkAuthorize/allow_update_pods/status-8         	  500000	      2826 ns/op	    1632 B/op	      27 allocs/op
BenchmarkAuthorize/forbid_educate_dolphins-8          	  500000	      2667 ns/op	    1632 B/op	      27 allocs/op
PASS
```

with this PR:
```
$ go test ./plugin/pkg/auth/authorizer/rbac/ -run foo -bench . -benchmem
goos: darwin
goarch: amd64
pkg: k8s.io/kubernetes/plugin/pkg/auth/authorizer/rbac
BenchmarkAuthorize/allow_list_pods-8         	  500000	      2622 ns/op	    1632 B/op	      27 allocs/op
BenchmarkAuthorize/allow_update_pods/status-8         	  500000	      2767 ns/op	    1632 B/op	      27 allocs/op
BenchmarkAuthorize/forbid_educate_dolphins-8          	  500000	      2611 ns/op	    1632 B/op	      27 allocs/op
PASS
```

would like to see evidence this improves something before doing it