@simo5 2018-01-18
@liggitt this was inspired by a comment of yours downstream: https://github.com/openshift/origin/pull/13259#discussion_r105028421
If we think this is not really goign to provide value upstream or downstream I can close this PR, just let me know.